resolving package lock issue
