+++ src/backend/middleware/verify-jira-middleware/index.ts
@@ -1,15 +1,36 @@
-import jwt from "atlassian-jwt";
+import {Installation} from "../../models";
+import {NextFunction, Request, Response} from "express";
+import {TokenType, verifySymmetricJwtTokenMiddleware} from "../../../common/jwt";
+import {decodeSymmetric, getAlgorithm} from "atlassian-jwt";
+import { booleanFlag, BooleanFlags } from "../../../config/feature-flags";
 
-import { Installation } from "../../models";
-import { NextFunction, Request, Response } from "express";
+const verifyTokenIsDecodableMiddleware = (sharedSecret: string, req: Request, next: NextFunction) => {
+	try {
+		const token = req.session.jwt || req.body?.token || req.query.jwt;
 
-export default async (
+		if(!token) {
+			next(new Error("Unauthorized"));
+			return
+		}
+
+		// The JWT contains a `qsh` field that can be used to verify
+		// the request body / query
+		// See https://bitbucket.org/atlassian/atlassian-connect-express/src/f434e5a9379a41213acf53b9c2689ce5eec55e21/lib/middleware/authentication.js?at=master&fileviewer=file-view-default#authentication.js-227
+		decodeSymmetric(token, sharedSecret, getAlgorithm(token));
+
+		next();
+	} catch (error) {
+		next(new Error("Unauthorized"));
+	}
+
+}
+
+const verifyJiraJwtMiddleware = (tokenType: TokenType) => async (
 	req: Request,
 	res: Response,
 	next: NextFunction
 ): Promise<void> => {
 	const jiraHost = req.session.jiraHost || req.body?.jiraHost;
-	const token = req.session.jwt || req.body?.token;
 	const installation = await Installation.getForHost(jiraHost);
 
 	if (!installation) {
@@ -23,14 +44,14 @@ export default async (
 			installation.clientKey && `${installation.clientKey.substr(0, 5)}***`
 	});
 
-	try {
-		// The JWT contains a `qsh` field that can be used to verify
-		// the request body / query
-		// See https://bitbucket.org/atlassian/atlassian-connect-express/src/f434e5a9379a41213acf53b9c2689ce5eec55e21/lib/middleware/authentication.js?at=master&fileviewer=file-view-default#authentication.js-227
-		jwt.decode(token, installation.sharedSecret);
-
-		next();
-	} catch (error) {
-		next(new Error("Unauthorized"));
+	if (await booleanFlag(BooleanFlags.FIX_IFRAME_ENDPOINTS_JWT, false, jiraHost)) {
+		verifySymmetricJwtTokenMiddleware(installation.sharedSecret, tokenType, req, res, next);
+	} else {
+		verifyTokenIsDecodableMiddleware(installation.sharedSecret, req, next);
 	}
+
 };
+
+export const verifyJiraJwtTokenMiddleware = verifyJiraJwtMiddleware(TokenType.normal);
+
+export const verifyJiraContextJwtTokenMiddleware = verifyJiraJwtMiddleware(TokenType.context);
